Refactor:  implement layered architecture for `get_gpu_info` (Linux Only)
