+++ src-tauri/src/_tests/utils/formatter_test.rs
@@ -67,6 +67,7 @@ mod tests {
   }
 
   #[test]
+  #[cfg(target_os = "windows")]
   fn test_rounded_kibibytes_display() {
     let kibibytes = RoundedKibibytes {
       kibibytes: Kibibytes(500),
@@ -88,6 +89,7 @@ mod tests {
   }
 
   #[test]
+  #[cfg(target_os = "windows")]
   fn test_celsius_to_fahrenheit() {
     let value = 100; // 100°C
     let result = format_temperature(
@@ -99,6 +101,7 @@ mod tests {
   }
 
   #[test]
+  #[cfg(target_os = "windows")]
   fn test_fahrenheit_to_celsius() {
     let value = 212; // 212°F
     let result = format_temperature(
@@ -110,6 +113,7 @@ mod tests {
   }
 
   #[test]
+  #[cfg(target_os = "windows")]
   fn test_celsius_to_fahrenheit_negative() {
     let value = -40; // -40°C
     let result = format_temperature(
@@ -121,6 +125,7 @@ mod tests {
   }
 
   #[test]
+  #[cfg(target_os = "windows")]
   fn test_fahrenheit_to_celsius_negative() {
     let value = -40; // -40°F
     let result = format_temperature(
@@ -132,6 +137,7 @@ mod tests {
   }
 
   #[test]
+  #[cfg(target_os = "windows")]
   fn test_no_conversion() {
     let value = 25; // 25°C
     let result = format_temperature(
